refactor: swap new cli commands for old
Replace the old commands with those that use the new-api directly.

Resolves #328.
