+++ packages/hardhat-plugin/src/type-extensions.ts
@@ -2,12 +2,12 @@
 import "hardhat/types/config";
 import "hardhat/types/runtime";
 
-import type { IgnitionWrapper } from "./ignition-wrapper";
 import type { IgnitionConfig } from "./index";
 
 import { IgnitionHelper } from "./ignition-helper";
 
 declare module "hardhat/types/config" {
+  // TODO: double check each of these makes sense
   export interface ProjectPathsUserConfig {
     ignition?: string;
     deployments?: string;
@@ -29,10 +29,6 @@ declare module "hardhat/types/config" {
 
 declare module "hardhat/types/runtime" {
   export interface HardhatRuntimeEnvironment {
-    ignition: IgnitionWrapper;
-  }
-
-  export interface HardhatRuntimeEnvironment {
-    ignition2: IgnitionHelper;
+    ignition: IgnitionHelper;
   }
 }
